+++ apps/docs/tailwind.config.ts
@@ -1,8 +1,8 @@
-import { sharedConfig, stone } from '@codefixlabs/tailwindcss';
+import { sharedConfig, slate } from '@codefixlabs/tailwindcss';
 import type { Config } from 'tailwindcss';
 
 const config: Pick<Config, 'presets' | 'plugins'> = {
-  plugins: [stone],
+  plugins: [slate],
   presets: [sharedConfig],
 };