+++ packages/tailwindcss/src/color/index.ts
@@ -1,1 +1,2 @@
 export * from '@/color/stone';
+export * from '@/color/slate';