+++ packages/ui/src/react/accordion.tsx
@@ -61,13 +61,14 @@ export const AccordionTrigger = forwardRef<
     };
   }
 >(({ children, className, classNames, ...props }, forwardedRef) => (
-  <Header className={twMerge('flex', classNames?.header)}>
+  <Header className={twMerge('flex', classNames?.header)} data-test-id="header">
     <Trigger
       className={twMerge(
         'group flex flex-1 cursor-pointer items-center justify-between py-4 text-base font-medium outline-none transition',
         className,
         classNames?.trigger,
       )}
+      data-test-id="trigger"
       ref={forwardedRef}
       {...props}
     >
@@ -79,6 +80,7 @@ export const AccordionTrigger = forwardRef<
             'group-data-state-open:rotate-180 text-accent-foreground size-4 transition-transform duration-300 ease-[cubic-bezier(0.87,_0,_0.13,_1)]',
             classNames?.icon,
           )}
+          data-test-id="icon"
         />
       </>
     </Trigger>