+++ packages/ui/src/react/input.tsx
@@ -147,6 +147,7 @@ export const Input = forwardRef<
           'text-muted-foreground absolute left-3 top-1/2 -translate-y-1/2',
           classNames?.startIcon,
         )}
+        data-test-id="start-icon"
       >
         {startIcon}
       </span>
@@ -160,6 +161,7 @@ export const Input = forwardRef<
           hasStartIcon: Boolean(startIcon),
         }),
       )}
+      data-test-id="input"
       ref={forwardedRef}
       {...props}
     />
@@ -170,6 +172,7 @@ export const Input = forwardRef<
           'text-muted-foreground absolute right-3 top-1/2 -translate-y-1/2',
           classNames?.endIcon,
         )}
+        data-test-id="end-icon"
       >
         {endIcon}
       </span>