+++ packages/ui/tailwind.config.ts
@@ -1,7 +1,8 @@
 import type { Config } from 'tailwindcss';
-import { sharedConfig } from '@codefixlabs/tailwindcss';
+import { sharedConfig, slate } from '@codefixlabs/tailwindcss';
 
-const config: Pick<Config, 'presets'> = {
+const config: Pick<Config, 'presets' | 'plugins'> = {
+  plugins: [slate],
   presets: [sharedConfig],
 };